requirements.txt
Django==1.11.4
djangorestframework==3.6.4
python-decouple==3.1
dj-database-url==0.4.2
dj-static==0.0.6
itsdangerous==0.24

dj-database-url==0.4.2
dj-static==0.0.6
static3==0.7.0
gunicorn==19.7.1
psycopg2==2.7.1
django-extensions==1.7.9

